client: don't enforce JRD content type
JRD documents should be returned with an application/jrd+json (or other
JSON) content-type, but even if they don't there's no real need to
enforce that.  At this point in the code path, we've already read the
entire response body, so we might as well attempt to parse it as JSON,
Postel's Law and all.